requirements.txt
httpx==0.26.0
orjson==3.9.10
pydantic==2.5.2
moviepy==1.0.3
pycryptodomex==3.14.1
zstandard==0.21.0
